pyproject.toml
[project]
name = "weixiaotien"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-markdown-table>=1.3.0",
    "pydantic-ai-slim[openai]>=0.1.11",
]
